refactor: rename useAuthorProfiles to useBlueskyAuthorProfiles for clarity
Add a comment to hopefully clarify how useBlueskyAuthorProfiles works with
the bluesky-author-profiles server end point.
